aiogram_bot/requirements.txt
aiogram 
asyncpg
psycopg2-binary
redis 
neo4j